Chore/bugs and fixes (#56)
* The game now tries a little harder to stop you from selling stocks you dont own

* format

* Selling a stock actually makes you lose it now.

* Quantity of stocks was a BigDecimal for some reason, it is now an int.

* Removed commented test stuff

* Advancing week now send you to todaymenu

* Fixed some more bigdecimal quantities

* Ngl i did this with claude, this quantity bigdecimal stuff was not worth it

* removed some garbage

* Alot of stuff regarding challenges. Reworked challenges slightly and  created the ProgressMenu to display them.

* Some tests broke when changing challenge stuff

* Fixed some tests doing stupid stuff

* Removed mentioning of watchlist, that feature is getting cut i think.

* Made the todaymenu show change in net worth from week to week instead of since start.

* Made the application save automatically when the application is closed without saving.

* Fixed bug in todaymenu

* test stuff idk

* Input when buying and selling is now validated

* Some method name changes for parity and some general code improvements

* You can now search for stocks, yippie

* Removed some cut stuff from exchangemenu

* We do not condone gambling.

+++ src/main/java/ntnu/gruppe21/model/Exchange.java
@@ -4,7 +4,6 @@ package ntnu.gruppe21.model;
 
 import static java.util.stream.Collectors.toList;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
@@ -170,7 +169,7 @@ public class Exchange {
    * @param player The player doing the purchase.
    * @return The transaction.
    */
-  public Transaction buy(String symbol, BigDecimal quantity, Player player) {
+  public Transaction buy(String symbol, int quantity, Player player) {
     Stock stock = getStock(symbol);
     Share share = new Share(stock, quantity, stock.getSalesPrice());
     Transaction purchase = transactionFactory.createPurchase(share, week);
@@ -189,7 +188,7 @@ public class Exchange {
    * @param player The player doing the sale.
    * @return The transaction.
    */
-  public Transaction sell(String symbol, BigDecimal quantity, Player player) {
+  public Transaction sell(String symbol, int quantity, Player player) {
     Stock stock = getStock(symbol);
     Share share = new Share(stock, quantity, stock.getSalesPrice());
     return transactionFactory.createSale(share, week);
